Use int32_t for position output

+++ src/tools/icn2img.cpp
@@ -105,7 +105,8 @@ int main( int argc, char ** argv )
             dstfile += ".png";
             shortdstfile += ".png";
 #endif
-            std::cout << "Image " << ii + 1 << " has offset of [" << head.offsetX << ", " << head.offsetY << "]" << std::endl;
+            std::cout << "Image " << ii + 1 << " has offset of [" << static_cast<int32_t>( head.offsetX ) << ", " << static_cast<int32_t>( head.offsetY ) << "]"
+                      << std::endl;
 
             fheroes2::Save( image, dstfile, 23 );
         }
